ai/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Data & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
tqdm>=4.65.0

# Evaluation
scikit-learn>=1.2.0

# Web Application
gradio>=4.0.0

# Notebook (EDA)
ipykernel>=6.0.0
ipywidgets>=8.0.0
